Dao - fix core data Articles
int64 -> int16

+++ True-X/Source/Articles/Articles.h
@@ -2,7 +2,7 @@
 //  Articles.h
 //  True-X
 //
-//  Created by Dao Nguyen on 6/2/13.
+//  Created by InfoNam on 6/3/13.
 //  Copyright (c) 2013 Dao Nguyen. All rights reserved.
 //
 
@@ -12,7 +12,7 @@
 
 @interface Articles : NSManagedObject
 
-@property (nonatomic) int64_t att1_id;
+@property (nonatomic) int16_t att1_id;
 @property (nonatomic, retain) NSString * att2_title;
 @property (nonatomic, retain) NSString * att3_thumbnailURL;
 @property (nonatomic, retain) NSString * att4_descriptionText;
